+++ snapshots/public/totem-web/lib/client/pub_totem_web_api.dart
@@ -1,0 +1,22 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+import 'package:degenerate_runtime/degenerate_runtime.dart';
+import '../apis/default_api.dart';
+import '../apis/events_api.dart';
+import '../apis/spaces_api.dart';
+
+/// Root SDK client providing access to all API groups.
+///
+/// ```dart
+/// final sdk = PubTotemWebApi(ApiConfig(client: myClient));
+/// sdk.$default.totemApiCurrentUser();
+/// ```
+final class PubTotemWebApi {
+  final ApiConfig _config;
+
+  PubTotemWebApi(this._config);
+
+  late final DefaultApi $default = DefaultApi(_config);
+  late final EventsApi events = EventsApi(_config);
+  late final SpacesApi spaces = SpacesApi(_config);
+}